moved setPath to tools
