+++ main.py
@@ -6,7 +6,7 @@ import re
 import praw
 import requests
 
-from tools import argument, blue, green, options, red, show_splash, create_config
+from tools import argument, blue, green, options, red, show_splash, create_config, setPath
 
 VERSION = "0.1"
 
@@ -123,23 +123,6 @@ class R3dditScrapper:
         else:
             # restart the program if not using terminal arguments
             main()
-
-
-def setPath():
-    print(blue("Enter the path you want to save to"))
-    """Set the path to download to"""
-    config = configparser.ConfigParser()
-    config.read("config.ini")
-    # if path exists, use it
-    if not config.has_section(section="Path"):
-        config.add_section("Path")
-    config.set("Path", "path", input("Enter the path to download to: "))
-    # check if path exists
-    if not os.path.exists(config["Path"]["path"]):
-            print(red("Path does not exist"))
-            setPath()
-    with open("config.ini", "w") as f:
-        config.write(f)
 
 
 def getInput():
